+++ SO/Labs/Lab5/src/SchedulerRR.java
@@ -14,7 +14,7 @@ class SchedulerRR extends Scheduler
 	public ArrayList<TimeSlot> schedule ()
 	{
 		ArrayList<TimeSlot> tsList = new ArrayList<TimeSlot> ();
-		double timer = 0;
+		int timer = 0;
 
 		for (Process proc : this.pList) {
 			TimeSlot newSlot = new TimeSlot (proc, timer, timer + proc.getBurstTime());
